+++ backend/src/main/java/com/carffeine/carffeine/admin/service/AdminCityService.java
@@ -6,12 +6,12 @@ import com.carffeine.carffeine.admin.service.dto.CityCreateRequest;
 import com.carffeine.carffeine.admin.service.dto.CityUpdateRequest;
 import com.carffeine.carffeine.member.domain.Member;
 import com.carffeine.carffeine.member.domain.MemberRepository;
-import com.carffeine.carffeine.station.domain.city.City;
-import com.carffeine.carffeine.station.domain.city.CityRepository;
+import com.carffeine.carffeine.city.domain.City;
+import com.carffeine.carffeine.city.domain.CityRepository;
 import com.carffeine.carffeine.station.domain.station.Latitude;
 import com.carffeine.carffeine.station.domain.station.Longitude;
-import com.carffeine.carffeine.station.exception.city.CityException;
-import com.carffeine.carffeine.station.exception.city.CityExceptionType;
+import com.carffeine.carffeine.city.exception.CityException;
+import com.carffeine.carffeine.city.exception.CityExceptionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
