+++ backend/src/main/java/com/carffeine/carffeine/city/exception/CityException.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.station.exception.city;
+package com.carffeine.carffeine.city.exception;
 
 import com.carffeine.carffeine.common.exception.BaseException;
 import com.carffeine.carffeine.common.exception.ExceptionType;
