+++ backend/src/main/java/com/carffeine/carffeine/city/infrastructure/repository/CityQueryRepository.java
@@ -1,13 +1,13 @@
-package com.carffeine.carffeine.station.infrastructure.repository.city;
+package com.carffeine.carffeine.city.infrastructure.repository;
 
-import com.carffeine.carffeine.station.infrastructure.repository.city.dto.CitySearchResponse;
+import com.carffeine.carffeine.city.infrastructure.repository.dto.CitySearchResponse;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-import static com.carffeine.carffeine.station.domain.city.QCity.city;
+import static com.carffeine.carffeine.city.domain.QCity.city;
 import static com.querydsl.core.types.Projections.constructor;
 
 @RequiredArgsConstructor
