+++ backend/src/main/java/com/carffeine/carffeine/city/service/CityRequester.java
@@ -1,0 +1,10 @@
+package com.carffeine.carffeine.city.service;
+
+import com.carffeine.carffeine.city.domain.City;
+
+import java.util.List;
+
+public interface CityRequester {
+
+    List<City> requestCities();
+}
