+++ backend/src/test/java/com/carffeine/carffeine/admin/integration/AdminCityIntegrationTest.java
@@ -5,8 +5,8 @@ import com.carffeine.carffeine.helper.integration.IntegrationTest;
 import com.carffeine.carffeine.member.domain.Member;
 import com.carffeine.carffeine.member.domain.MemberRepository;
 import com.carffeine.carffeine.member.domain.MemberRole;
-import com.carffeine.carffeine.station.domain.city.City;
-import com.carffeine.carffeine.station.domain.city.CityRepository;
+import com.carffeine.carffeine.city.domain.City;
+import com.carffeine.carffeine.city.domain.CityRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
