+++ backend/src/test/java/com/carffeine/carffeine/admin/integration/AdminCityIntegrationTestFixture.java
@@ -4,8 +4,7 @@ import com.carffeine.carffeine.admin.common.CustomPage;
 import com.carffeine.carffeine.admin.controller.dto.CityResponse;
 import com.carffeine.carffeine.admin.service.dto.CityCreateRequest;
 import com.carffeine.carffeine.admin.service.dto.CityUpdateRequest;
-import com.carffeine.carffeine.helper.integration.IntegrationTest;
-import com.carffeine.carffeine.station.domain.city.City;
+import com.carffeine.carffeine.city.domain.City;
 import io.restassured.RestAssured;
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.http.ContentType;
