+++ backend/src/test/java/com/carffeine/carffeine/station/domain/city/CityCustomRepositoryImplTest.java
@@ -1,7 +1,8 @@
 package com.carffeine.carffeine.station.domain.city;
 
+import com.carffeine.carffeine.city.domain.City;
 import com.carffeine.carffeine.helper.integration.IntegrationTest;
-import com.carffeine.carffeine.station.infrastructure.repository.city.CityCustomRepository;
+import com.carffeine.carffeine.city.infrastructure.repository.CityCustomRepository;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
